Extract BucketStorage interfaces into separate source files.
