+++ packages/service-core/src/storage/ReplicationLock.ts
@@ -1,0 +1,5 @@
+export interface ReplicationLock {
+  sync_rules_id: number;
+
+  release(): Promise<void>;
+}
