+++ packages/service-core/src/storage/storage-index.ts
@@ -7,3 +7,8 @@ export * from './SourceTable.js';
 export * from './StorageEngine.js';
 export * from './StorageProvider.js';
 export * from './WriteCheckpointAPI.js';
+export * from './BucketStorageFactory.js';
+export * from './BucketStorageBatch.js';
+export * from './SyncRulesBucketStorage.js';
+export * from './PersistedSyncRulesContent.js';
+export * from './ReplicationLock.js';
